Select safer MTD when escalating TO already-full dose
The case dealt with here is when current-cohort tally
indicates escalation to an already-full higher dose.
In this case, my earlier rule of simply declaring
whatever dose we first fail to enroll (because full)
effectively 'forgets the lesson' that higher dose
already taught us. You would be recommending an RP2D
whose latest tally indicated de-escalation!

Furthermore, I believe this change also (as if safety
weren't enough!) renders the RP2D PATH-INDEPENDENT,
given the design. That is, the final recommendation
can be reconstructed from the list of final tallies,
together with the final decision for each dose.
This is an especially desirable property for CCDs,
since it accentuates the sole dependence on tallies.